Rearrange modules and functions to avoid circular imports and clean up
